+++ secureProject/src/main/java/secure/project/secureProject/configuration/WebConfig.java
@@ -15,7 +15,7 @@ public class WebConfig implements WebMvcConfigurer {
     public void addCorsMappings(final CorsRegistry registry){
         registry.addMapping("/**")
                 .allowedOriginPatterns("*")
-                .allowedMethods("GET","POST","PUT","DELETE","HEAD","OPTIONS")
+                .allowedMethods("PATCH","GET","POST","PUT","DELETE","HEAD","OPTIONS")
                 .allowedHeaders("*")
                 .allowCredentials(true);
     }
