+++ secureProject/src/main/java/secure/project/secureProject/controller/LoginController.java
@@ -5,12 +5,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import secure.project.secureProject.dto.reqeust.SignInReqeustDto;
+import secure.project.secureProject.dto.request.SignInRequestDto;
 import secure.project.secureProject.dto.response.ResponseDto;
-import secure.project.secureProject.dto.response.TokenDto;
 import secure.project.secureProject.security.LoginService;
 import secure.project.secureProject.security.jwt.JwtToken;
-import secure.project.secureProject.util.SecurityUtil;
 
 @RestController
 @RequestMapping("/login")
@@ -20,14 +18,9 @@ public class LoginController {
 
     @PostMapping("/sign-in")
     public ResponseDto<JwtToken> signIn(
-            @RequestBody SignInReqeustDto signInReqeustDto
+            @RequestBody SignInRequestDto signInRequestDto
             ) {
-        System.err.println("controller까지는 들어옴");
-        return new ResponseDto<>(loginService.signIn(signInReqeustDto));
+        return new ResponseDto<>(loginService.signIn(signInRequestDto));
     }
 
-    @PostMapping("/test")
-    public ResponseDto<String> test() {
-        return new ResponseDto<>(SecurityUtil.getCurrentUsername());
-    }
 }
