+++ secureProject/src/main/java/secure/project/secureProject/dto/request/BasketAddItemRequestDto.java
@@ -1,4 +1,4 @@
-package secure.project.secureProject.dto.reqeust;
+package secure.project.secureProject.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
