+++ secureProject/src/main/java/secure/project/secureProject/dto/request/ItemAddAmountRequestDto.java
@@ -1,4 +1,4 @@
-package secure.project.secureProject.dto.reqeust;
+package secure.project.secureProject.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,8 +7,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class ItemReqeustDto {
-    private String itemName;
+public class ItemAddAmountRequestDto {
+    private Long itemId;
     private Integer itemAmount;
-    private Integer itemPrice;
 }
