+++ secureProject/src/main/java/secure/project/secureProject/dto/request/ItemRequestDto.java
@@ -1,4 +1,4 @@
-package secure.project.secureProject.dto.reqeust;
+package secure.project.secureProject.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,7 +7,8 @@ import lombok.NoArgsConstructor;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class ItemAddAmountRequestDto {
-    private Long itemId;
+public class ItemRequestDto {
+    private String itemName;
     private Integer itemAmount;
+    private Integer itemPrice;
 }
