+++ secureProject/src/main/java/secure/project/secureProject/dto/request/SignInRequestDto.java
@@ -1,4 +1,4 @@
-package secure.project.secureProject.dto.reqeust;
+package secure.project.secureProject.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,10 +7,8 @@ import lombok.NoArgsConstructor;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class UserReqeustDto {
-    private Long id;
+public class SignInRequestDto {
     private String loginId;
     private String password;
-    private String nickname;
-    private Long point;
 }
+
