+++ secureProject/src/main/java/secure/project/secureProject/dto/request/UserIdRequestDto.java
@@ -1,4 +1,4 @@
-package secure.project.secureProject.dto.reqeust;
+package secure.project.secureProject.dto.request;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -7,6 +7,6 @@ import lombok.Setter;
 @Getter
 @Setter
 @NoArgsConstructor
-public class UserIdReqeustDto {
+public class UserIdRequestDto {
     private Long userId;
 }
