+++ secureProject/src/main/java/secure/project/secureProject/dto/request/UserRequestDto.java
@@ -1,4 +1,4 @@
-package secure.project.secureProject.dto.reqeust;
+package secure.project.secureProject.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,8 +7,10 @@ import lombok.NoArgsConstructor;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class SignInReqeustDto {
+public class UserRequestDto {
+    private Long id;
     private String loginId;
     private String password;
+    private String nickname;
+    private Long point;
 }
-
