+++ secureProject/src/main/java/secure/project/secureProject/repository/AdminItemRepository.java
@@ -7,7 +7,6 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import secure.project.secureProject.domain.Item;
 
-import javax.swing.text.html.Option;
 import java.util.Optional;
 
 public interface AdminItemRepository extends JpaRepository<Item, Long> {
