+++ secureProject/src/main/java/secure/project/secureProject/repository/BasketRepository.java
@@ -1,16 +1,13 @@
 package secure.project.secureProject.repository;
 
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import secure.project.secureProject.domain.Basket;
 import secure.project.secureProject.domain.Item;
 import secure.project.secureProject.domain.User;
-import secure.project.secureProject.dto.response.BasketDto;
 
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 
 public interface BasketRepository extends JpaRepository<Basket, Long> {
