+++ secureProject/src/main/java/secure/project/secureProject/repository/CustomerItemRepository.java
@@ -6,13 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import secure.project.secureProject.domain.Item;
-import secure.project.secureProject.domain.Order;
-import secure.project.secureProject.domain.OrderItem;
 
-import javax.swing.text.html.Option;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
 
 public interface CustomerItemRepository extends JpaRepository<Item, Long> {
         @Query("SELECT i FROM Item i WHERE (:itemName IS NULL OR i.itemName = :itemName)")
