+++ secureProject/src/main/java/secure/project/secureProject/repository/OrderItemRepository.java
@@ -5,7 +5,6 @@ import secure.project.secureProject.domain.Order;
 import secure.project.secureProject.domain.OrderItem;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface OrderItemRepository extends JpaRepository<OrderItem, Long> {
     List<OrderItem> findByOrder(Order order);
