+++ secureProject/src/main/java/secure/project/secureProject/repository/OrderRepository.java
@@ -1,12 +1,9 @@
 package secure.project.secureProject.repository;
 
-import org.aspectj.weaver.ast.Or;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-import secure.project.secureProject.domain.Item;
 import secure.project.secureProject.domain.Order;
 import secure.project.secureProject.domain.User;
