+++ secureProject/src/main/java/secure/project/secureProject/repository/UserRepository.java
@@ -11,19 +11,6 @@ import java.util.Optional;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
-//    Optional<User> findByLoginId(String loginId);
-//
-//    @Query(value = "SELECT u.id AS id, u.role AS userRole FROM User u WHERE u.id = :userId")
-//    Optional<UserLoginForm> findUserForAuthentication(@Param("userId") Long userId);
-//
-//    @Query("SELECT u.id AS id, u.role AS userRole FROM User u WHERE u.id = :userId AND u.isLogin = true AND u.refreshToken = :refreshToken")
-//    Optional<UserLoginForm> findByIdAndRefreshToken(@Param("userId") Long userId, @Param("refreshToken") String refreshToken);
-//
-//
-//    public interface UserLoginForm {
-//        Long getId();
-//        UserRole getUserRole();
-//    }
     Optional<User> findByNickname(String nickname);
     Optional<User> findByLoginId(String loginId);
 }
