+++ secureProject/src/main/java/secure/project/secureProject/security/LoginService.java
@@ -5,15 +5,10 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import secure.project.secureProject.domain.User;
-import secure.project.secureProject.dto.reqeust.SignInReqeustDto;
-import secure.project.secureProject.enums.UserRole;
+import secure.project.secureProject.dto.request.SignInRequestDto;
 import secure.project.secureProject.exception.ApiException;
 import secure.project.secureProject.exception.ErrorDefine;
 import secure.project.secureProject.repository.UserRepository;
@@ -29,25 +24,24 @@ public class LoginService {
     private final AuthenticationManagerBuilder authenticationManagerBuilder;
     private final JwtTokenProvider jwtTokenProvider;
     @Transactional
-    public JwtToken signIn(SignInReqeustDto signInReqeustDto) {
+    public JwtToken signIn(SignInRequestDto signInReqeustDto) {
         // 1. username + password 를 기반으로 Authentication 객체 생성
         // 이때 authentication 은 인증 여부를 확인하는 authenticated 값이 false
         User userNickname = userRepository.findByLoginId(signInReqeustDto.getLoginId())
                         .orElseThrow(()-> new ApiException(ErrorDefine.USER_NOT_FOUND));
-        System.err.println("service 1");
+
         UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken(userNickname.getNickname(), signInReqeustDto.getPassword());
-        System.err.println("service 211");
+
         // 2. 실제 검증. authenticate() 메서드를 통해 요청된 Member 에 대한 검증 진행
         // authenticate 메서드가 실행될 때 CustomUserDetailsService 에서 만든 loadUserByUsername 메서드 실행
         Authentication authentication = authenticationManagerBuilder.getObject().authenticate(authenticationToken);
-        System.err.println("service 3"+ authentication);
 
         User user = userRepository.findByNickname(userNickname.getNickname())
                 .orElseThrow(() -> new ApiException(ErrorDefine.USER_NOT_FOUND));
 
         // 3. 인증 정보를 기반으로 JWT 토큰 생성
         JwtToken jwtToken = jwtTokenProvider.generateToken(authentication, user.getUserRole());
-        System.err.println(jwtToken.getAccessToken());
+
         return jwtToken;
     }
