+++ secureProject/src/main/java/secure/project/secureProject/security/jwt/JwtTokenProvider.java
@@ -43,7 +43,7 @@ public class JwtTokenProvider {
         long now = (new Date()).getTime();
 
         // Access Token 생성
-        Date accessTokenExpiresIn = new Date(now + 86400000);
+        Date accessTokenExpiresIn = new Date(now + 2592000000L);
         String accessToken = Jwts.builder()
                 .setSubject(authentication.getName())
                 .claim("auth", userRole) // 여기 수정
@@ -53,7 +53,7 @@ public class JwtTokenProvider {
 
         // Refresh Token 생성
         String refreshToken = Jwts.builder()
-                .setExpiration(new Date(now + 86400000))
+                .setExpiration(new Date(now + 2592000000L))
                 .signWith(key, SignatureAlgorithm.HS256)
                 .compact();
 
@@ -68,7 +68,7 @@ public class JwtTokenProvider {
     public Authentication getAuthentication(String accessToken) {
         // Jwt 토큰 복호화
         Claims claims = parseClaims(accessToken);
-        System.err.println(claims);
+
         if (claims.get("auth") == null) {
             throw new RuntimeException("권한 정보가 없는 토큰입니다.");
         }
