+++ secureProject/src/main/java/secure/project/secureProject/service/UserService.java
@@ -3,10 +3,6 @@ package secure.project.secureProject.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import secure.project.secureProject.domain.User;
-import secure.project.secureProject.dto.reqeust.UserReqeustDto;
-import secure.project.secureProject.exception.ApiException;
-import secure.project.secureProject.exception.ErrorDefine;
 import secure.project.secureProject.repository.UserRepository;
 
 @Service
