+++ secureProject/src/main/java/secure/project/secureProject/util/S3UploadUtil.java
@@ -1,6 +1,5 @@
 package secure.project.secureProject.util;
 
-
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
